Add BuildLLVM.bat script for building LLVM & creating lib/ folder
Had a problem with using a newer LLVM commit in the submodule but
using generated library & header files from an older commit of
LLVM (I'd built it a while ago, and thought I could be clever
and reuse the generated files instead of build it again... I'm suprised
it took this long to break).

The way this broke was that a commit had (in a later version than the
one I used to build the binaries/headers) added support for a new OpenMP
feature. The 'in tree' headers required new code in the generated headers
but since they were generated from a previous version, everything
massively broke in the most confusing way when trying to use a random class
(RecursiveASTVisitor...).

TL;DR this script should automate the process of configuring and
building LLVM and then copying the required files to the lib/
folder.