Add a distinct pointer type
Pointers are being modelled as opaque in this IR - i.e they don't have an underlying type.

This makes it a bit simpler to implement and having typed pointers doesn't actually solve that much
(in terms of safety you can still cast between pointer types).

This is inspired by LLVM which is moving towards opaque pointers as well
https://llvm.org/docs/OpaquePointers.html