Add better slot tracking support for debug printing
So this isn't perfect and slots won't always be the same, depending on what you print since
each Print(...) overload creates its own slot tracker for now.
For example, if you print a single instruction
    iadd 0:i32, 1:i32, %0:i32

the first named value (the result register) will be allocated to slot 0, since there have been no
other slots encountered.

On the other hand if you print the same instruction but part of the following basic block:
.0:
   isub 0:i32, 0:i32, %0:i32
   iadd 0:i32, 1:i32, %1:i32 # this is the same insn

it will be allocated to slot 1 since the first slot went to the first encountered value (the result from the isub).

This isn't great and really slots should be fixed for each vreg, but from the current most common usage
this works well enough

(the same problem applies to printing either singular basic blocks or basic blocks as part of a funciton etc...)