Add support for array subscripting
E.g a[10] etc...

This involves adding a lea instruction which is a bit like getelementptr from LLVM, but simpler.

Basically lea takes 3 (4 including the type) operands: a base type, an input pointer, an index/offset (in terms of the given base type)
and an output pointer that'll store the calculated address.

Example:

stack_alloc [i32 x 1], %0:ptr
lea [i32], %0:ptr, 0:i32, %1:ptr

This will calculate the address of the first element in the array (allocated into %0)