Simplify/codense lvalue code generation
E.g. there were a few times where we handled DeclRefExpr seperately,
but these could all be condensed into one function that handles lvalues
(since DelcRefExpr's are variable names, which are lvalues). We can then
rely on the LValueToRValue implicit cast that clang injects to convert
from the lvalue (which is a pointer to the variable in this ir) to the actual value (e.g. an i32 if the variable is an integer)