Fixes to sizeof & integral casts
Sizeof was broken due to using getArgumentType() instead of getTypeOfArgument().
The latter will get the type of the sizeof param even if the param is an expression,
whereas getArgumentType() only works if the argument is an immedate type name (asserts otherwise).

The integral cast pointer are as follows:
 - IntegerType that was supported to represent the target type was actually representing the source type (typo...)
 - Return a new constant int instead of using SetType when casting a constant int - this is because ConstantInts
are pooled, and using SetType would change the type of all ConstantInts in the program for that number.
 -  Fix calculating maximum value for any given bit width - a 1 constant in a bitwise operation was implicitly an int
but needed to be dealing with values that are potentially 64 bit.

These fixes get the passing tests for c-testsuite up from 54 -> 60.