Add initial enum implementation
Eh, this turned out really really messy.

Basically ended up implementing a constant evaluator from a clang::Expr - this came up
from the fact that we can't reply EnumConstantDecl::getInitVal() all the time, for example
when using sizeof or other target specific expressions in the enum init expression.
Clang would use its own target details, but that might not be what we want. Trying to get clang
to have the right target was a pain, so  this is where I ended up...

It's pretty awful.