Add initial support for global (file scope) variables
This is not perfect yet and required quite a bit of reworking:
Certain instructions were taking only VirtualRegisterName as operands, this was
meant to signify "output" operands, and would prevent you from passing a ConstantInt*
as an output of add etc...
However, we also want global variables (which are pointers) to be used in this "output" context
as well, so everything that used VirtualRegisterName in that way has now been changed to just use
the base Value.

Global variables can have initialisers, but support for this isn't great, e.g. you couldn't use
an expression (even a constant one) since we don't have generic constant folding yet (only can
use initialisers that evaluate to a ConstantInt).

Still need the concept of linkage in there as well, e.g. static (internal) vs external linkage.

Added some tests, but nothing comprehensive yet.