Legalise struct initialisers to the appropriate lfa/store pairs
For example,

    store { 123:i32, 541:i32 }, %0:ptr

becomes

  lfa [ST:0], %0:ptr, %1:ptr
  store 123:i32, %1:ptr

  lfa [ST:1], %0:ptr, %2:ptr
  store 123:i32, %2:ptr

e.g. a memberwise store.

In addition, now genlegal will iteratively run until it's
legalised all it can, since legalising certain expressions might
produce more expressions that need to be legalised.

For example, nested struct initialisers or having arrays of structs
which are both in place initialised.

This runs the risk of an infinite loop, but I think that's quite unlikely
atm, but can always add a max iterations timeout if it comes up.