Add some temporary scripts
hcx.bat - This is a placeholder for a compiler driver, it is meant to take assembly emitted by the compiler and
compile it into an actual binary. This should be done by the driver process eventually.

run-arm.sh - This is a utility script to just run the generated executable in a qemu userspace VM. Needs
to be run from a properly configured linux environment (using WSL for convenience)