Fix ICE from frontend emitting a basic block for non empty void functions
Void functions that have a conditional as the last statement would
cause the frontend to emit a tail block for the conditional statement
but since there was no other code after it it would be left empty
when processing of the function had finished.
There was some code that was meant to detect this and inject a return
(either returning an undefined value or nothing if it was a void function)
but that code was buggy.

This adds a load more tests to detect this in the future, since the ICE
was only caught when trying to add another feature.