Add stack_alloc hoisting to genlegal
This is described in `docs\[Transform] Hoist Stack Allocations.md`.
TL;DR it's not really a nessesary optimisation since stack_alloc is a "fake" instruction
that just reserves stack space for the compiler and doesn't get executed as such.
This just makes sure all stack allocs are in the first basic block, which makes later
analysis/operations easier to deal with.