Add ValidationPass, a pass that checks the IR is well formed
If it's not any errors are logged to the `validate` channel and an ICE is triggered.

Currently this is first run on the IR immediately produced by the frontend, and then
it is ran after every pass to ensure the pass hasn't botched anything too terribly.

Running after every pass is probably a bit extreme, it probably makes more sense to
(by default) run once after the frontend and once just before final code generation,
and then can hide the other "per pass" checks behind a debug/developer flag.

Still a couple of instructions that should have rules added to them, but got
most of them I think.

Turns out most code is already well formed, had to fix up cconv pass so that it changes
the function type to void when it removes the return values from all return instructions
but bar that everything else was fine after a bit of tweaking.

c-testsuite compilation pass rate went from 96 -> 93 so obviously there is something fishy
going on somewhere but thats an investigation for later.