Fix a few bugs with size_t being the wrong size
This was caused by a mismatch between what clang thought the target was, what I was treating
size_t as in the backend, and me forgetting to change an i64 -> i32 in the frontend.

Clang now should parse for a ARMv7 Raspberry Pi like architecture, and this should match what
we expect (before it was parsing for Intel x64, so size-t == 8 bytes and not 4 etc...). This should stop
any other problems like this in the future.

Still got to implement returning larger than native word size values by output parameter, but that can come... :)