Load return value directly into r0
We can now take advantage of the fact that return values are stored on the stack
(from the return combine pass), and instead of loading the return value from
memory into a temporary vreg and returning it, we can just load the value
straight into r0.

This needs some tests, and breaks one (trying to return a struct, we need output parameter returns
for that...) but mostly seems ok.