Add NULL, size_t, ptrdiff_t and wchar_t definitions to stddef.h
This gets LibC setup in cmake, and passes through a temporary include directory (for development
purposes) to clang.
(This adds a cmake target for LibC, even though it shouldn't be used to build it, eventually we'll need
to use the hxc compiler itself to build libc, but for now we've got include stuff only).