Testify: Start compiling files to assembly and trying to execute them
Now Testify (for the Helix testsuite anyway) will try to compile the input files to assembly.
If that is ok it will try to assemble and link them, and proving *that* works, then try to execute
the generated executable using qemu-user (via a invoked WSL instance).

This breaks everything.

This whole system is sort of thrown together, but it seems to work?

There is a massive drop in the helix testsuite pass rate though :D (got a screenshot for the report)