Make -o work for assembly output files & executable files
If -S is set -o is for assembly output file name, else we pass it directly
to GCC and it chooses what to do.

If -o isn't specified then we let GCC use it's own default (a.out most likely), but it probably
makes sense to fix this and always use our own defaults if nessesary.