Physical registers can now be 8, 16 or 32 bits wide
Note not 64 bit, as registers on the ARM are 32 bits wide so 64 bit will require doing arithmetic
on multiple registers.

This is so we can emit the load & store instructions for the correct type, as previously if the code
tried to store a byte in a memory address the compiler would try to store the value into 4 bytes
and not just the singular byte.

This goes some way to addressing #10, but still need to handle the signedness problem, e.g.
emitting ldrs[b|h]/strs[b|h] for signed bytes & shorts.