Make spill order deterministic (based on order of first use)
Was having determinism problems where the order of spills was different accross
runs causing different runs of the compiler to place stack allocated variables at different
places in the stack frame.
This was due to emitting the stack_allocs (the order of which
is used to determine offset in the stack frame) from an unordered_map that they where
stored in (with the key being a pointer...).

Now before emitting stack_allocs for the spills they are sorted into a determistic order
(could be better? not that bothered atm since this register allocator is temporary).

The RunN.bat script is just a utility for running the compiler N times and comparing
outputs against previous (far from perfect). This is just a scratch script & only in source
control for a backup.