Clean up load & store instruction patterns
This adds support for specifying * (astrisk) as the type in
`match_operand`, to mean "any type".

This is a convenience,  saves duplicating a load of patterns for 8/16/32 bit variants.

This patch also moves the load register <-> register & store register <-> register handling
into custom code, to save a bit of duplicating & allow some flexibility.