Fix bug where if an insn spilled multiple operands, they would overwrite each other
Originally the "spill register" (scratch register used for storing all
spilled operands whilst they are stored to/loaded from the stack) was
r0. This was just for simplicity (laziness) and I was hoping that
no instruction ever needed to spill multiple operands (that was really
stupid, no assert to check that or anything...).

This fixes that, allowing the use of r[0-3] for spilling (allthough in
practice I don't think r3 itself will ever be used, worst case with
the standard 3 operand instructions and if all are spilled only r[0-2]
will be needed.

Found this whilst investigating slowdown in binary_search benchmark,
after the optimisations that were added to make strlen faster.

Turned out that mem2reg was inadvertedly causing this bug, see
notes in "Optimising binary search"