Add InsnSeq
InsnSeq is just a sequence of instructions, using the
existing intrusive list infrastructure.

Primary use case for this is building sequences of
instructions and injecting them into an existing block.

(previously the only instruction collection was a basic block
and it doesn't make sense to create a basic block as a temporary
container).

Unit tests included :-)